+++ tests/FactoryTest.php
@@ -10,11 +10,14 @@ use Innmind\OperatingSystem\{
 };
 use Innmind\TimeContinuum\Clock;
 use Innmind\Filesystem\{
+    Adapter\Filesystem,
     File,
     File\Content,
     Directory,
+    CaseSensitivity,
 };
 use Innmind\Url\Path;
+use Innmind\Immutable\Attempt;
 use Symfony\Component\Filesystem\Filesystem as FS;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -48,7 +51,16 @@ class FactoryTest extends TestCase
         $path = \sys_get_temp_dir().'/innmind/filesystem/';
         (new FS)->remove($path);
 
-        $os = Factory::build(Config::of()->caseInsensitiveFilesystem());
+        $os = Factory::build(
+            Config::of()->mountFilesystemVia(
+                static fn($path, $config) => Attempt::of(
+                    static fn() => Filesystem::mount(
+                        $path,
+                        $config->io(),
+                    )->withCaseSensitivity(CaseSensitivity::insensitive),
+                ),
+            ),
+        );
         $adapter = $os
             ->filesystem()
             ->mount(Path::of($path))